Add files via upload
This includes the Lab Session on How To Build a Machine Learning Classifier in Python with Scikit-learn. In this tutorial, I implemented a simple machine-learning algorithm in Python using Scikit-learn, a machine-learning tool for Python. Using a database of breast cancer tumor information, I used a Naive Bayes (NB) classifier that predicts whether or not a tumor is malignant or benign.